RR-901 - Rename HighestEducationLevel to EducationLevel
