+++ domain/learningandworkprogress/src/main/kotlin/uk/gov/justice/digital/hmpps/domain/learningandworkprogress/induction/QualificationLevel.kt
@@ -1,0 +1,16 @@
+package uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction
+
+/**
+ * An enumeration of qualification levels in line with the [gov.uk qualification levels](https://www.gov.uk/what-different-qualification-levels-mean/list-of-qualification-levels).
+ */
+enum class QualificationLevel {
+  ENTRY_LEVEL,
+  LEVEL_1,
+  LEVEL_2,
+  LEVEL_3,
+  LEVEL_4,
+  LEVEL_5,
+  LEVEL_6,
+  LEVEL_7,
+  LEVEL_8,
+}
