+++ domain/learningandworkprogress/src/main/kotlin/uk/gov/justice/digital/hmpps/domain/learningandworkprogress/induction/dto/CreatePreviousQualificationsDto.kt
@@ -1,11 +1,11 @@
 package uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction.dto
 
-import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction.HighestEducationLevel
+import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction.EducationLevel
 import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction.Qualification
 
 data class CreatePreviousQualificationsDto(
   val prisonNumber: String,
-  val educationLevel: HighestEducationLevel,
+  val educationLevel: EducationLevel,
   val qualifications: List<Qualification>,
   val prisonId: String,
 )
