+++ domain/learningandworkprogress/src/testFixtures/kotlin/uk/gov/justice/digital/hmpps/domain/learningandworkprogress/induction/PreviousQualificationsBuilder.kt
@@ -1,12 +1,12 @@
 package uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction
 
-import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction.HighestEducationLevel.SECONDARY_SCHOOL_LEFT_BEFORE_TAKING_EXAMS
+import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction.EducationLevel.SECONDARY_SCHOOL_LEFT_BEFORE_TAKING_EXAMS
 import java.time.Instant
 import java.util.UUID
 
 fun aValidPreviousQualifications(
   reference: UUID = UUID.randomUUID(),
-  educationLevel: HighestEducationLevel? = SECONDARY_SCHOOL_LEFT_BEFORE_TAKING_EXAMS,
+  educationLevel: EducationLevel? = SECONDARY_SCHOOL_LEFT_BEFORE_TAKING_EXAMS,
   qualifications: List<Qualification> = listOf(aValidQualification()),
   createdBy: String? = "asmith_gen",
   createdByDisplayName: String? = "Alex Smith",
