+++ domain/learningandworkprogress/src/testFixtures/kotlin/uk/gov/justice/digital/hmpps/domain/learningandworkprogress/induction/dto/PreviousQualificationsDtoBuilder.kt
@@ -1,14 +1,14 @@
 package uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction.dto
 
-import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction.HighestEducationLevel
-import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction.HighestEducationLevel.SECONDARY_SCHOOL_LEFT_BEFORE_TAKING_EXAMS
+import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction.EducationLevel
+import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction.EducationLevel.SECONDARY_SCHOOL_LEFT_BEFORE_TAKING_EXAMS
 import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction.Qualification
 import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.induction.aValidQualification
 import java.util.UUID
 
 fun aValidCreatePreviousQualificationsDto(
   prisonNumber: String = "A1234AB",
-  educationLevel: HighestEducationLevel = SECONDARY_SCHOOL_LEFT_BEFORE_TAKING_EXAMS,
+  educationLevel: EducationLevel = SECONDARY_SCHOOL_LEFT_BEFORE_TAKING_EXAMS,
   qualifications: List<Qualification> = listOf(aValidQualification()),
   prisonId: String = "BXI",
 ) =
@@ -21,7 +21,7 @@ fun aValidCreatePreviousQualificationsDto(
 
 fun aValidUpdatePreviousQualificationsDto(
   reference: UUID = UUID.randomUUID(),
-  educationLevel: HighestEducationLevel = SECONDARY_SCHOOL_LEFT_BEFORE_TAKING_EXAMS,
+  educationLevel: EducationLevel = SECONDARY_SCHOOL_LEFT_BEFORE_TAKING_EXAMS,
   qualifications: List<Qualification> = listOf(aValidQualification()),
   prisonId: String = "BXI",
 ) =
