+++ src/testFixtures/kotlin/uk/gov/justice/digital/hmpps/educationandworkplanapi/app/database/jpa/entity/induction/PreviousQualificationsEntityBuilder.kt
@@ -8,7 +8,7 @@ fun aValidPreviousQualificationsEntity(
   id: UUID? = null,
   reference: UUID = UUID.randomUUID(),
   prisonNumber: String = aValidPrisonNumber(),
-  educationLevel: HighestEducationLevel = HighestEducationLevel.SECONDARY_SCHOOL_TOOK_EXAMS,
+  educationLevel: EducationLevel = EducationLevel.SECONDARY_SCHOOL_TOOK_EXAMS,
   qualifications: MutableList<QualificationEntity> = mutableListOf(aValidQualificationEntity()),
   createdAt: Instant? = null,
   createdAtPrison: String = "BXI",
@@ -39,7 +39,7 @@ fun aValidPreviousQualificationsEntityWithJpaFieldsPopulated(
   id: UUID? = UUID.randomUUID(),
   reference: UUID = UUID.randomUUID(),
   prisonNumber: String = aValidPrisonNumber(),
-  educationLevel: HighestEducationLevel = HighestEducationLevel.SECONDARY_SCHOOL_TOOK_EXAMS,
+  educationLevel: EducationLevel = EducationLevel.SECONDARY_SCHOOL_TOOK_EXAMS,
   qualifications: MutableList<QualificationEntity> = mutableListOf(aValidQualificationEntity()),
   createdAt: Instant? = Instant.now(),
   createdAtPrison: String = "BXI",
